Address comments #2: Remove singular selection view.
 - Combine *SliceSelectionView.java and *SlicesSelectionView.java into
 one, do selection count checking at selection view composite creation
 time, and set the composite layout accordingly.
 - This can help reduce file numbers and make it easier to locate
 selection view composite.
