--- gapic/src/main/com/google/gapid/perfetto/views/FrameEventsMultiSelectionView.java
@@ -1,77 +1,0 @@
-/*
- * Copyright (C) 2019 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.gapid.perfetto.views;
-
-import static com.google.gapid.perfetto.TimeSpan.timeToString;
-import static com.google.gapid.widgets.Widgets.createTreeColumn;
-import static com.google.gapid.widgets.Widgets.createTreeViewer;
-import static com.google.gapid.widgets.Widgets.packColumns;
-
-import com.google.gapid.perfetto.models.FrameEventsTrack;
-
-import org.eclipse.jface.viewers.ITreeContentProvider;
-import org.eclipse.jface.viewers.LabelProvider;
-import org.eclipse.jface.viewers.TreeViewer;
-import org.eclipse.swt.SWT;
-import org.eclipse.swt.layout.FillLayout;
-import org.eclipse.swt.widgets.Composite;
-
-/**
- * Displays information about a list of selected slices.
- */
-public class FrameEventsMultiSelectionView extends Composite {
-  public FrameEventsMultiSelectionView(Composite parent, FrameEventsTrack.Slices slices) {
-    super(parent, SWT.NONE);
-    setLayout(new FillLayout());
-
-    FrameEventsTrack.Node[] nodes = FrameEventsTrack.organizeSlicesToNodes(slices);
-
-    TreeViewer viewer = createTreeViewer(this, SWT.NONE);
-    viewer.getTree().setHeaderVisible(true);
-    viewer.setContentProvider(new ITreeContentProvider() {
-      @Override
-      public Object[] getElements(Object inputElement) {
-        return nodes;
-      }
-
-      @Override
-      public boolean hasChildren(Object element) {
-        return false;
-      }
-
-      @Override
-      public Object getParent(Object element) {
-        return null;
-      }
-
-      @Override
-      public Object[] getChildren(Object element) {
-        return null;
-      }
-    });
-    viewer.setLabelProvider(new LabelProvider());
-
-    createTreeColumn(viewer, "Name", e -> n(e).name);
-    createTreeColumn(viewer, "Self Time", e -> timeToString(n(e).self));
-    createTreeColumn(viewer, "Layers", e -> String.join(", ", n(e).layers));
-    viewer.setInput(slices);
-    packColumns(viewer.getTree());
-  }
-
-  protected static FrameEventsTrack.Node n(Object o) {
-    return (FrameEventsTrack.Node)o;
-  }
-}
