+++ gapic/src/main/com/google/gapid/perfetto/views/ThreadStateSlicesSelectionView.java
@@ -16,10 +16,16 @@
 package com.google.gapid.perfetto.views;
 
 import static com.google.gapid.perfetto.TimeSpan.timeToString;
+import static com.google.gapid.widgets.Widgets.createBoldLabel;
+import static com.google.gapid.widgets.Widgets.createLabel;
 import static com.google.gapid.widgets.Widgets.createTableColumn;
 import static com.google.gapid.widgets.Widgets.createTableViewer;
 import static com.google.gapid.widgets.Widgets.packColumns;
+import static com.google.gapid.widgets.Widgets.withLayoutData;
+import static com.google.gapid.widgets.Widgets.withSpans;
 
+import com.google.gapid.perfetto.models.ProcessInfo;
+import com.google.gapid.perfetto.models.ThreadInfo;
 import com.google.gapid.perfetto.models.ThreadTrack;
 
 import org.eclipse.jface.viewers.IStructuredContentProvider;
@@ -27,6 +33,8 @@ import org.eclipse.jface.viewers.LabelProvider;
 import org.eclipse.jface.viewers.TableViewer;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.FillLayout;
+import org.eclipse.swt.layout.GridData;
+import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 
 /**
@@ -34,8 +42,42 @@ import org.eclipse.swt.widgets.Composite;
  */
 public class ThreadStateSlicesSelectionView extends Composite {
   public ThreadStateSlicesSelectionView(
-      Composite parent, ThreadTrack.StateSlices slices) {
+      Composite parent, State state, ThreadTrack.StateSlices slices) {
     super(parent, SWT.NONE);
+    if (slices.count == 1) {
+      setSingleSliceView(state, slices);
+    } else if (slices.count > 1) {
+      setMultiSlicesView(slices);
+    }
+  }
+
+  private void setSingleSliceView(State state, ThreadTrack.StateSlices slice) {
+    setLayout(new GridLayout(2, false));
+    withLayoutData(createBoldLabel(this, "Slice:"), withSpans(new GridData(), 2, 1));
+
+    createLabel(this, "Time:");
+    createLabel(this, timeToString(slice.times.get(0) - state.getTraceTime().start));
+
+    createLabel(this, "Duration:");
+    createLabel(this, timeToString(slice.durs.get(0)));
+
+    ThreadInfo thread = state.getThreadInfo(slice.utids.get(0));
+    if (thread != null) {
+      ProcessInfo process = state.getProcessInfo(thread.upid);
+      if (process != null) {
+        createLabel(this, "Process:");
+        createLabel(this, process.getDisplay());
+      }
+
+      createLabel(this, "Thread:");
+      createLabel(this, thread.getDisplay());
+    }
+
+    createLabel(this, "State:");
+    createLabel(this, slice.states.get(0).label);
+  }
+
+  private void setMultiSlicesView(ThreadTrack.StateSlices slices) {
     setLayout(new FillLayout());
 
     ThreadTrack.Entry[] entries = ThreadTrack.organizeSlicesToEntry(slices);
